+++ Debug/Object-Builder.sh
@@ -1,3 +1,0 @@
-#This will make a dump of the objects for the exe
-#use this with the stackdump to figure out what went wrong.
-objdump -D -S ../PlatformerExperiment-PRIVATE.exe > build.rasm.log