+++ Source/config.cpp
@@ -1,6 +1,5 @@
 /*****************************************************************************/
 #include "config.h"
-#include "globals.h"
 #include "version.h"
 /*****************************************************************************/
 /////////////////////////////////////////////////
@@ -12,6 +11,7 @@ clsConfig::clsConfig() {
   /////////////////////////////////////////////////
   /// @brief This is the default constructor for the Config Class, set values as
   ///        the following defaults before loaded from the file.
+  ///        * Human Build = false;
   ///        * Logging = true
   ///        * Show Map = true
   ///        * Append Time = true
@@ -29,6 +29,7 @@ clsConfig::clsConfig() {
   /////////////////////////////////////////////////
 
 	//Set the values as some default value.
+	values.blnHumanBuild = false;
 	values.blnLogging = true;
 	values.blnShowMap = true;
 	values.blnAppendTime = true;
@@ -43,7 +44,7 @@ clsConfig::clsConfig() {
 	values.uintMaxPlayerSteps = 500;
 	values.uintPlayersPerGen = 40;
 	values.uintNumberOfBreed = 10;
-	if (global::blnDebugMode) {printf("Config Constructor called.\n");}
+	if (global::blnDebugMode) { printf("Config Constructor called.\n"); }
 }
 /*****************************************************************************/
 bool clsConfig::exists(void) {
@@ -54,8 +55,11 @@ bool clsConfig::exists(void) {
 
 	//Returns true or false if config file exists
 	FILE* pTempFile = fopen(FileName, "r");
-	if (pTempFile == NULL) {return false;}
-	else {return true;}
+	if (pTempFile == NULL) { return false; }
+	else {
+    fclose(pTempFile);
+    return true;
+  }
 }
 /*****************************************************************************/
 void clsConfig::make(void) {
@@ -72,14 +76,17 @@ void clsConfig::make(void) {
 	fprintf(configFile,"Config File for the program.\n");
 	fprintf(configFile,"%s\n",DEFINED_VER_FULLVERSION_STRING);
 
-    //Write the config with the default values defined when the object was created.
+  //Write the config with the default values defined when the object was created.
+  fprintf(configFile,"Human Build: %u\n", values.blnHumanBuild ? 1 : 0);
 	fprintf(configFile,"First Generation Steps: %u\n",values.uintFirstGen);
 	fprintf(configFile,"Generation Increase: %u\n", values.uintGenIncrease);
 	fprintf(configFile,"Gens Past Growth: %u\n", values.uintGensPastGrowth);
   fprintf(configFile,"Max Player steps: %u\n", values.uintMaxPlayerSteps);
   fprintf(configFile,"Players per Generation: %u\n", values.uintPlayersPerGen);
-  fprintf(configFile,"Best players per generation: %u\n", values.uintNumberOfBreed);
-	fprintf(configFile,"Percent Mutation Chance: %u\n", values.uintMutationChance);
+  fprintf(configFile,"Best players per generation: %u\n",
+      values.uintNumberOfBreed);
+	fprintf(configFile,"Percent Mutation Chance: %u\n",
+      values.uintMutationChance);
 	fprintf(configFile,"Log to File: %u\n", values.blnLogging ? 1 : 0);
 	fprintf(configFile,"Hard mode: %u\n", values.blnHardMode ? 1 : 0);
 	fprintf(configFile,"Show map on update: %u\n", values.blnShowMap ? 1 : 0);
@@ -106,11 +113,15 @@ char clsConfig::verisonCheck(const char *ConfigVerison) {
 
 	uint C_MajorNum, C_MinorNum, C_PatchNum;
 	char C_SoftwareStatus[15], P_SoftwareStatus[15];
-	sscanf(ConfigVerison,"%u.%u.%u-%s",&C_MajorNum,&C_MinorNum,&C_PatchNum,&C_SoftwareStatus);
+	sscanf(ConfigVerison,"%u.%u.%u-%s",
+      &C_MajorNum,&C_MinorNum,&C_PatchNum,&C_SoftwareStatus);
 	sscanf(DEFINED_VER_FULLVERSION_STRING,"%*u.%*u.%*u-%s", &P_SoftwareStatus);
 	if (global::blnDebugMode) {
-    printf("Config: v %u %u %u %s\n",C_MajorNum,C_MinorNum,C_PatchNum,C_SoftwareStatus);
-    printf("Program: v %u %u %u %s\n",DEFINED_VER_MAJOR, DEFINED_VER_MINOR, DEFINED_VER_PATCH, P_SoftwareStatus);
+    printf("Config: v %u %u %u %s\n",
+        C_MajorNum,C_MinorNum,C_PatchNum,C_SoftwareStatus);
+    printf("Program: v %u %u %u %s\n",
+        DEFINED_VER_MAJOR, DEFINED_VER_MINOR, DEFINED_VER_PATCH,
+        P_SoftwareStatus);
   }
 
 
@@ -142,29 +153,49 @@ void clsConfig::load(void) {
 
 	char chrTempString[50];
 	int intTempBool, intValuesScanned;
+
+  //Get if there is logging
+	fgets(chrTempString,50,configFile);
+	intValuesScanned = sscanf(chrTempString, "%*s %*s %d", &intTempBool);
+	if (intValuesScanned < 1) { printf("ERROR!"); intTempBool = 0; }
+	if(global::blnDebugMode) { printf("Human Build \t \t \t %d\n",intTempBool); }
+	values.blnHumanBuild = (intTempBool == 1);
+
 	//Get the First Generation Input from config
 	fgets(chrTempString,50,configFile);
-	intValuesScanned = sscanf(chrTempString,"%*s %*s %*s %d",&values.uintFirstGen);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintFirstGen = 100;}
-	if(global::blnDebugMode) {printf("First Gen Steps \t \t %2d\n",values.uintFirstGen);}
+	intValuesScanned = sscanf(chrTempString,"%*s %*s %*s %d",
+                           &values.uintFirstGen);
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintFirstGen = 100; }
+	if(global::blnDebugMode) {
+    printf("First Gen Steps \t \t %2d\n",values.uintFirstGen);
+  }
 
 	//Get Amount the inputs increase by each generation
 	fgets(chrTempString,50,configFile);
-	intValuesScanned = sscanf(chrTempString, "%*s %*s %d",&values.uintGenIncrease);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintGenIncrease = 100;}
-	if(global::blnDebugMode) {printf("Generation Increase \t \t %2d\n",values.uintGenIncrease);}
+	intValuesScanned = sscanf(chrTempString, "%*s %*s %d",
+                           &values.uintGenIncrease);
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintGenIncrease = 100; }
+	if(global::blnDebugMode) {
+    printf("Generation Increase \t \t %2d\n",values.uintGenIncrease);
+  }
 
 	//Get the number of generations past growth
 	fgets(chrTempString,50,configFile);
-	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %d",&values.uintGensPastGrowth);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintGensPastGrowth = 10;}
-	if(global::blnDebugMode) {printf("Gens Past Growth \t \t %u\n",values.uintGensPastGrowth);}
+	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %d",
+                           &values.uintGensPastGrowth);
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintGensPastGrowth = 10; }
+	if(global::blnDebugMode) {
+    printf("Gens Past Growth \t \t %u\n",values.uintGensPastGrowth);
+  }
 
   //Get the max number of steps
 	fgets(chrTempString,50,configFile);
-	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %d",&values.uintMaxPlayerSteps);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintMaxPlayerSteps = 500;}
-	if(global::blnDebugMode) {printf("Max steps \t \t \t %u\n",values.uintMaxPlayerSteps);}
+	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %d",
+                           &values.uintMaxPlayerSteps);
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintMaxPlayerSteps = 500; }
+	if(global::blnDebugMode) {
+    printf("Max steps \t \t \t %u\n",values.uintMaxPlayerSteps);
+  }
 	if(values.uintMaxPlayerSteps < values.uintFirstGen) {
     printf("ERROR! Max steps set to value lower than initial.\nReseting value.\n");
     getchar();
@@ -174,14 +205,19 @@ void clsConfig::load(void) {
   //Get the number of players per gen
 	fgets(chrTempString,50,configFile);
 	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %d",&values.uintPlayersPerGen);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintPlayersPerGen = 40;}
-	if(global::blnDebugMode) {printf("Player per gen \t \t \t %u\n",values.uintPlayersPerGen);}
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintPlayersPerGen = 40; }
+	if(global::blnDebugMode) {
+    printf("Player per gen \t \t \t %u\n",values.uintPlayersPerGen);
+  }
 
 		//Get the number of best players per generation
 	fgets(chrTempString,50,configFile);
-	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %*s %d",&values.uintNumberOfBreed);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintNumberOfBreed = 10;}
-	if(global::blnDebugMode) {printf("Best players per gen \t \t %u\n",values.uintNumberOfBreed);}
+	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %*s %d",
+                           &values.uintNumberOfBreed);
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintNumberOfBreed = 10; }
+	if(global::blnDebugMode) {
+    printf("Best players per gen \t \t %u\n",values.uintNumberOfBreed);
+  }
 	if(values.uintNumberOfBreed > values.uintPlayersPerGen) {
     printf("ERROR! Number of best players per generation higher than number of players.\nReseting value.\n");
     getchar();
@@ -191,43 +227,48 @@ void clsConfig::load(void) {
 	//Get the percent chance of mutation
 	//Should I make this a float?
 	fgets(chrTempString,50,configFile);
-	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %u", &values.uintMutationChance);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintMutationChance = 15;}
-	if(global::blnDebugMode) {printf("Percent Mutation Chance \t %u\n",values.uintMutationChance);}
+	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %u",
+                           &values.uintMutationChance);
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintMutationChance = 15; }
+	if(global::blnDebugMode) {
+    printf("Percent Mutation Chance \t %u\n",values.uintMutationChance);
+  }
 
 	//Get if there is logging
 	fgets(chrTempString,50,configFile);
 	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %d", &intTempBool);
-	if (intValuesScanned < 1) {printf("ERROR!"); intTempBool = 1;}
-	if(global::blnDebugMode) {printf("Log to file \t \t \t %d\n",intTempBool);}
+	if (intValuesScanned < 1) { printf("ERROR!"); intTempBool = 1; }
+	if(global::blnDebugMode) { printf("Log to file \t \t \t %d\n",intTempBool); }
 	values.blnLogging = (intTempBool == 1);
 
 	//Check if hard mode is enabled.
 	fgets(chrTempString,50,configFile);
 	intValuesScanned = sscanf(chrTempString,"%*s %*s %d",&intTempBool);
-	if (intValuesScanned < 1) {printf("ERROR!"); intTempBool = 0;}
-	if(global::blnDebugMode) {printf("Hard mode \t \t \t %u\n",intTempBool);}
-	if(intTempBool == 1) {values.blnHardMode = true;}
-	else {values.blnHardMode = false;}
+	if (intValuesScanned < 1) { printf("ERROR!"); intTempBool = 0; }
+	if(global::blnDebugMode) { printf("Hard mode \t \t \t %u\n",intTempBool); }
+	if(intTempBool == 1) { values.blnHardMode = true; }
+	else { values.blnHardMode = false; }
 
 	//Check if show map is enabled
 	fgets(chrTempString,50,configFile);
 	intValuesScanned = sscanf(chrTempString, "%*s %*s %*s %*s %d",&intTempBool);
-	if (intValuesScanned < 1) {printf("ERROR!"); intTempBool = 0;}
-	if(global::blnDebugMode) {printf("Show Map Update \t \t %d\n",intTempBool);}
+	if (intValuesScanned < 1) { printf("ERROR!"); intTempBool = 0; }
+	if(global::blnDebugMode) { printf("Show Map Update \t \t %d\n",intTempBool); }
 	values.blnShowMap = (intTempBool == 1);
 
 	//Get the seed used in the config
 	fgets(chrTempString,50,configFile);
 	intValuesScanned = sscanf(chrTempString,"%*s %*s [%u]", &values.uintSeed);
-	if (intValuesScanned < 1) {printf("ERROR!"); values.uintSeed = 0;}
-	if(global::blnDebugMode) {printf("Random Seed \t \t \t %d\n",values.uintSeed);}
+	if (intValuesScanned < 1) { printf("ERROR!"); values.uintSeed = 0; }
+	if(global::blnDebugMode) {
+    printf("Random Seed \t \t \t %d\n",values.uintSeed);
+  }
 
 	//Check if append time
 	fgets(chrTempString,50,configFile);
 	intValuesScanned = sscanf(chrTempString,"%*s %*s %d",&intTempBool);
-	if (intValuesScanned < 1) {printf("ERROR!"); intTempBool = 1;}
-	if(global::blnDebugMode) {printf("Append Time \t \t \t %u\n",intTempBool);}
+	if (intValuesScanned < 1) { printf("ERROR!"); intTempBool = 1; }
+	if(global::blnDebugMode) { printf("Append Time \t \t \t %u\n",intTempBool); }
 	values.blnAppendTime = (intTempBool == 1);
 
 	//Get blank line explaining screen sizes
@@ -237,15 +278,21 @@ void clsConfig::load(void) {
 
 	//Get Screen Height
 	fgets(chrTempString,50,configFile);
-  intValuesScanned = sscanf(chrTempString, "%*s %*s %u", &values.uintScreenHeight);
-  if (intValuesScanned < 1) {printf("ERROR!"); values.uintScreenHeight = 0;}
-  if (global::blnDebugMode) {printf("Screen Height \t \t \t %u\n", values.uintScreenHeight);}
+  intValuesScanned = sscanf(chrTempString, "%*s %*s %u",
+                            &values.uintScreenHeight);
+  if (intValuesScanned < 1) { printf("ERROR!"); values.uintScreenHeight = 0; }
+  if (global::blnDebugMode) {
+    printf("Screen Height \t \t \t %u\n", values.uintScreenHeight);
+  }
 
   //Get Screen Width
   fgets(chrTempString,50,configFile);
-  intValuesScanned = sscanf(chrTempString, "%*s %*s %u", &values.uintScreenWidth);
-  if (intValuesScanned < 1) {printf("ERROR!"); values.uintScreenWidth = 0;}
-  if (global::blnDebugMode) {printf("Screen Width \t \t \t %u\n", values.uintScreenWidth);}
+  intValuesScanned = sscanf(chrTempString, "%*s %*s %u",
+                            &values.uintScreenWidth);
+  if (intValuesScanned < 1) { printf("ERROR!"); values.uintScreenWidth = 0; }
+  if (global::blnDebugMode) {
+    printf("Screen Width \t \t \t %u\n", values.uintScreenWidth);
+  }
 
 	fclose(configFile);
 	printf("\n\n");
@@ -333,51 +380,54 @@ uint clsConfig::getvalues(uchar Spot) {
 	//Returns just one value from the config.
 	//Useful when I don't need ALL the values.
 	switch (Spot) {
-  case cnfgLogging :
-    return values.blnLogging ? 1 : 0;
-    break;
-  case cnfgShowMap :
-    return values.blnShowMap ? 1 : 0;
-    break;
-  case cnfgAppendTime :
-    return values.blnAppendTime ? 1 : 0;
-    break;
-  case cnfgHardMode :
-    return values.blnHardMode ? 1 : 0;
-    break;
-  case cnfgFirstGen :
-    return values.uintFirstGen;
-    break;
-  case cnfgGenIncrease :
-    return values.uintGenIncrease;
-    break;
-  case cnfgGensPastGrowth :
-    return values.uintGensPastGrowth;
-    break;
-  case cnfgMutationChance :
-    return values.uintMutationChance;
-    break;
-  case cnfgSeed :
-    return values.uintSeed;
-    break;
-  case cnfgScreenHeight :
-    return values.uintScreenHeight;
-    break;
-  case cnfgScreenWidth :
-    return values.uintScreenWidth;
-    break;
-  case cnfgMaxSteps:
-    return values.uintMaxPlayerSteps;
-    break;
-  case cnfgPlayerGen:
-    return values.uintPlayersPerGen;
-    break;
-  case cnfgPlayerBreed:
-    return values.uintNumberOfBreed;
-    break;
-  default :
-    return 9999;
-    break;
+    case cnfgLogging :
+      return values.blnLogging ? 1 : 0;
+      break;
+    case cnfgShowMap :
+      return values.blnShowMap ? 1 : 0;
+      break;
+    case cnfgAppendTime :
+      return values.blnAppendTime ? 1 : 0;
+      break;
+    case cnfgHardMode :
+      return values.blnHardMode ? 1 : 0;
+      break;
+    case cnfgFirstGen :
+      return values.uintFirstGen;
+      break;
+    case cnfgGenIncrease :
+      return values.uintGenIncrease;
+      break;
+    case cnfgGensPastGrowth :
+      return values.uintGensPastGrowth;
+      break;
+    case cnfgMutationChance :
+      return values.uintMutationChance;
+      break;
+    case cnfgSeed :
+      return values.uintSeed;
+      break;
+    case cnfgScreenHeight :
+      return values.uintScreenHeight;
+      break;
+    case cnfgScreenWidth :
+      return values.uintScreenWidth;
+      break;
+    case cnfgMaxSteps:
+      return values.uintMaxPlayerSteps;
+      break;
+    case cnfgPlayerGen:
+      return values.uintPlayersPerGen;
+      break;
+    case cnfgPlayerBreed:
+      return values.uintNumberOfBreed;
+      break;
+    case cnfgHumanBuild:
+      return values.blnHumanBuild ? 1 : 0;
+      break;
+    default :
+      return 9999;
+      break;
 	}; //end switch
 
 	return 9999;