+++ Source/tick.h
@@ -4,7 +4,7 @@
 #include <time.h>
 #include <math.h>
 /*****************************************************************************/
-#include "main.h"
+#include "globals.h"
 /*****************************************************************************/
 /////////////////////////////////////////////////
 /// @class clsTick tick.h "source/tick.h"
@@ -14,36 +14,21 @@
 
 class clsTick {
   public:
-  //Default Constructor
-  clsTick();
+    clsTick();
 
-  //Functions
-  void wait(void); //Call this to activate the wait
-  void resetClock(void);
-  void startTimer(void);
-
-  void decClock(void); //This is so the clock will decrement even when show map is false.
-  uint getClockTime(void);
+    void wait(void); //Call this to activate the wait
+    void resetClock(void);
+    void startTimer(void);
+    void decClock(void); //This is so the clock will decrement even when show map is false.
+    uint getClockTime(void);
 
   private:
-  //Members
-  /** The Frame per second we want to see. */
-  uchar uchrFPS;
-
-  /** How long (in milliseconds) the program should wait for. */
-  ulong ulngSleepTime;
-
-  /** The time the player has left to complete the map (in seconds). */
-  uint uClock;
-
-  /** The clock as a floating point, will be converted to uint. */
-  float fClock;
-
-  /** The start of the timer since the last pause. */
-  clock_t timerStart;
-
-  /** The start of the clock. */
-  clock_t clockStart;
+    uchar uchrFPS; /**< The Frame per second we want to see. */
+    ulong ulngSleepTime; /**< How long (in milliseconds) the program should wait for. */
+    uint uClock; /**< The time the player has left to complete the map (in seconds). */
+    float fClock; /**< The clock as a floating point, will be converted to uint. */
+    clock_t timerStart; /**< The start of the timer since the last pause. */
+    clock_t clockStart; /**< The start of the clock. */
 };
 /*****************************************************************************/
 #endif